Feature/fix user data

+++ app/src/main/java/com/hapataka/questwalk/data/dto/UserDTO.kt
@@ -4,7 +4,7 @@ data class UserDTO(
     val id: String = "",
     var nickName: String = "",
     var characterId: Int = -1,
-    var totalTime: Int = 0,
+    var totalTime: Long = 0L,
     var totalDistance: Float = 0f,
-    var totalStep: Long = 0,
+    var totalStep: Long = 0L,
 )
